picture-ui/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
Pillow
